Clear templated NextJS web app
