+++ frontend/app/page.tsx
@@ -1,0 +1,8 @@
+export default function Page() {
+  return (
+    <div>
+      <h1 className="text-primary-dark text-4xl font-bold">Hello Header 1</h1>
+      <p className="text-accent text-base">Hello from Sample Page!</p>
+    </div>
+  );
+}
