+++ frontend/package.json
@@ -22,7 +22,7 @@
     "eslint": "^9",
     "eslint-config-next": "15.4.5",
     "postcss": "^8.5.6",
-    "tailwindcss": "3.4",
+    "tailwindcss": "^3.4.0",
     "typescript": "^5"
   }
 }
